vkid & network actions (#41)
* add cu, list

* add time field to medrecord

* need add auth

* add getting of vk access token

* add getting of vk access token

* receive from vk posts

* add save and get of vk access token by pref datastore

* fix errorUiState & nav in BottomBar & statusBar color

* add petswallscreen, viewmodel; relocate everything from loginscreen

need to fix an interface

* fix errorUiState & nav in BottomBar & statusBar color

* fix errorUiState & nav in BottomBar & statusBar color

* fix procedures data show

* fix some warnings

* fix detekt errors

* done except validations

* nothing changed

* change interface

* fix pet delete

* add medRecord Service

* add pull to refresh in list procedures

* rework it with API Handler

* fix ui after pet create

* fix pet and procedure deletions

* done interface

* added share func

* fix detekt

* add titles in NetworkRepository

* add update of local db for procs and medrecords

* fix Procedure Create

* fix context, properties, ImmutableList

* add Procedure Update & need to fix refresh titles after Procedure create

* add Procedure Update & need to fix refresh titles after Procedure create

* rework Procedure list screen

* add MedRecords list refresh

* add MedRecord Update and create

* fix errors

* fix detekt errors

* try to fix notifications

* change text in message

* remove selectable dates &

---------

+++ app/src/main/java/com/f4/mypet/ui/screens/procedure/show/ProcedureViewModel.kt
@@ -7,8 +7,9 @@ import com.f4.mypet.data.db.entities.Frequency
 import com.f4.mypet.data.db.entities.Procedure
 import com.f4.mypet.data.db.entities.ProcedureTitle
 import com.f4.mypet.data.db.entities.ProcedureType
+import com.f4.mypet.data.network.NetworkRepository
+import com.f4.mypet.data.network.model.NetworkResult
 import com.f4.mypet.ui.screens.procedure.FrequencyOptions
-import com.f4.mypet.util.PetDateTimeFormatter
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -19,51 +20,48 @@ import javax.inject.Inject
 
 @HiltViewModel
 class ProcedureViewModel @Inject constructor(
-    private val repository: Repository
+    private val repository: Repository,
+    private val networkRepository: NetworkRepository
 ) : ViewModel() {
     private val _procedureUiState = MutableStateFlow(
         Procedure(
-            0, 0, "",0,
-            LocalDateTime.parse("01.01.1001 00:00", PetDateTimeFormatter.dateTime),
-            "", LocalDateTime.parse("01.01.1001 00:00", PetDateTimeFormatter.dateTime),
-            0, 0, 0
+            0, 0, "", 0,
+            LocalDateTime.now().withMinute(0),
+            "", LocalDateTime.now().withMinute(0),
+            0, 0, -1
         )
     )
-    var title = ProcedureTitle(
-        name = "Неизвестно",
-        type = -1,
-        id = -1
-    )
-    var type = ProcedureType(
-        name = "Неизвестно",
-        id = title.type
-    )
-    var frequency = Frequency(
-        "Никогда",
-        "0"
-    )
-
     val procedureUiState = _procedureUiState.asStateFlow()
+    // если null, то ошибок при удалении не было
+    private val _msg = MutableStateFlow<String?>("")
+    val msg = _msg.asStateFlow()
+
+    var title = ProcedureTitle(name = "", type = -1, id = -1)
+    var type = ProcedureType(name = "", id = title.type)
+    var frequency = Frequency(option = "Никогда", frequency = "0")
 
     fun getProcedure(procedureId: Int) {
         viewModelScope.launch(Dispatchers.IO) {
             _procedureUiState.value = repository.getProcedure(procedureId)
-            title =
-                repository.getProcedureTitlesForCU().find { it.id == _procedureUiState.value.title }
-                    ?: title
-            type = repository.getProcedureTypes().find { it.id == title.type }
-                ?: type
             frequency = repository.getFrequency(_procedureUiState.value.frequencyOption)
+            title = repository.getProcedureTitlesForCU()
+                .find { it.id == _procedureUiState.value.title } ?: title
+            type = repository.getProcedureTypes()
+                .find { it.id == title.type } ?: type
 
             when (frequency.option) {
                 FrequencyOptions.Minutes.period -> frequency.frequency =
                     _procedureUiState.value.frequency + FrequencyOptions.Minutes.abbreviation
+
                 FrequencyOptions.Hours.period -> frequency.frequency =
                     _procedureUiState.value.frequency + FrequencyOptions.Hours.abbreviation
+
                 FrequencyOptions.Days.period -> frequency.frequency =
                     _procedureUiState.value.frequency + FrequencyOptions.Days.abbreviation
+
                 FrequencyOptions.Weeks.period -> frequency.frequency =
                     _procedureUiState.value.frequency + FrequencyOptions.Weeks.abbreviation
+
                 else -> frequency.frequency = FrequencyOptions.Never.abbreviation
             }
         }
@@ -71,8 +69,18 @@ class ProcedureViewModel @Inject constructor(
 
     fun deleteProcedure(procedure: Procedure) {
         viewModelScope.launch(Dispatchers.IO) {
-            //TODO add delete of frequency
-            repository.deleteProcedure(procedure)
+             when (val response = networkRepository.removeProcedure(procedure.id)) {
+                is NetworkResult.Success -> {
+                    _msg.value = null
+                    repository.deleteProcedure(procedure)
+                }
+                is NetworkResult.Error -> { _msg.value = response.msg }
+            }
         }
     }
+
+    fun resetMsg() {
+        _msg.value = ""
+    }
 }
+
